Add a utility class measurements for relative error analysis
